refactor: Changed the layout of the classes such that the required parameters are in the correct classes.
