+++ app/src/main/res/layout/location_view.xml
@@ -33,24 +33,20 @@
                 android:id="@+id/faveName"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@{location.name}"
-                android:hint="Location name"
+                android:text="Name"
                 android:layout_marginStart="105dp"
                 android:layout_marginTop="5dp"
                 android:textSize="15sp"
-                android:textColor="@color/white"
                 />
 
             <TextView
                 android:id="@+id/faveDescription"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@{location.description}"
-                android:hint="Description goes here"
+                android:text="Description"
                 android:layout_marginStart="105dp"
                 android:layout_marginTop="40dp"
                 android:textSize="12sp"
-                android:textColor="@color/white"
                 />
 
         </androidx.cardview.widget.CardView>
